requirements.txt
pyopengl == 20.0.1